fix(mmclient): Improve audio re-syncing
This attempts to sync audio to video using presentation timestamps, with
millisecond resolution. Only two operations are supported: playing silence to
let the video catch up, and skipping ahead to catch up to the video. An earlier
version of this patch experimented with playing audio at 105%/95% speed, but
that created a very negative user experience. That could be revisited in the
future using some algorithm which compensates for the changes in pitch.

This currently aims for 20-60ms of audio lag, which seems to strike a good
balance between stability (longer lag means more buffer) and subjective
latency.