fix(server): improve dmabuf juggling
The main fix here is to address a crash that resulted from reused dmabuf fds.
Since mesa (via Vulkan) owns the fd of a dmabuf once its passed in, we would
often end up double-closing a given fd once we dropped it on the smithay side.
If the fd we closed was presently being imported/reused in a new dmabuf, that
would cause the import to fail.