build: use separate crates without a workspace
The workspace feature means that dependencies are shared, and a single version
(with a single set of features) is locked across all workspace members. This
isn't useful for us, since we share almost no dependencies between the server
and client, and since latency-test depends on bevy, which has many, many, many
dependencies. It also causes problems with e.g. wayland-sys, which is forced to
pull in libwayland-client when building the server, for example.