fix(server): unlinearize when converting from RGB to YCbCr
Dmabufs, when they come to us from the application, are usually stored as sRGB,
with premultiplied alpha. To blend correctly, we need to convert them to linear.
We can't just specify _SRGB formats when importing, because wayland requires
premultiplied alpha in buffers, and vulkan's SRGB formats assume straight alpha.

However, fixing that uncovered a second problem, which is that we were doing the
YCbCr conversion on linear color values, rather than color values already in the
rec.709 color space. This happened to work because we were skipping a conversion
(to linear) on the import side and also skipping a conversion (to 709) on the
conversion side. However, it was only mostly correct; sRGB and 709 have the same
color primaries, but different transfer functions.