fix(client): do srgb-correct rendering
On most devices, we'll have swapchains expecting sRGB color, but we were
rendering to UNORM textures without doing sRGB conversion first - against the
spec. This only worked because of a second misunderstanding of the vulkan spec.
When a YUV texture is sampled with a VkSamplerYCbCrConversion, the resulting rgb
values are still in the color space they started; that is, vulkan does the
narrow/full expansion and the matrix multiplication, but doesn't apply any sort
of transfer function to linearize the rgb values, unlike with sRGB textures.

The correct thing to do is linearize using whatever transfer function the video
was tagged with, and then unlinearize using sRGB (or whatever the swapchain
expects).