feat(server): add enable_datagrams, off by default
This adds a configuration option to enable quic DATAGRAM support, and disables
it by default. Datagrams are better and the protocol is built around them, but
they're less robust in the face of packet loss, particularly without any FEC
mechanism.

We'll turn on datagram support by default once we have that.