fix(mmserver): don't discard buffers when resending frame callbacks
In this context, we have a configure which has not yet been acked, so we
want to discard the buffer and get the client to commit a new one.
However, in some situations, the previous buffer is actually valid for
our unacked commit, and GTK, at least, does not commit a new buffer in
that instance.

The fix is still to send a frame callback to tickle the client into
doing another commit(), but keep the buffer in case they want to ack and
commit without changing the buffer.