fix(mmserver): log when container setup fails
We redirect container output to a pipe, and poll it in the main
compositor loop. Unfortunately, that means that any output prior to a
failed call to `Container::spawn` never got captured, printed, or saved
to a bug report log.

This adds some code to pump the output pipe on failure, so that we can
see why the container setup failed.