feat!: implement forward error correction
Forward error correction, or FEC, allows us to spend bandwidth and a
little bit of CPU to increase the robustness of the video stream in the
face of lost packets. In this first simple implementation, we use an
inexpensive scheme called RaptorQ to protect the base layer of a video
stream (assuming hierarchical coding has created multiple layers).

This causes us to send 15% more data for those frames, with the
advantage that the frames are recoverable with up to 15% packets
missing.