feat!: implement forward error correction
Forward error correction, or FEC, allows us to spend bandwidth and a
little bit of CPU and latency to increase the robustness of the video
stream in the face of lost packets. In this first simple
implementation, we use an inexpensive scheme called RaptorQ, set at a
constant level of overhead for each video layer in the server config.