perf(mmserver): optimize the write path for frames
Any work we do preparing frames for sending increases latency, but it's
also just a lot of work - and doing that on the server thread seems to
gum up the works. On the other hand, both the audio and video encoders
have a spare thread just for writing out frames, with lots of downtime,
so with a bit of awkwardness we can push the work of doing the protobuf
encode and chunking out to those threads.