fix(mmserver): try to avoid colliding with the system x11 socket
A real xserver sometimes creates "abstract sockets", which transcend the
filesystem and can leak into our container (because we don't use a
network namespace). That means when running on a system with the system
xserver at :1, we would potentially end up colliding with it, and the app
would connect to the system xserver instead of xwayland.

Note that we can't just use DISPLAY=/path/to/internal/socket, because
that's not supported on even some recent libxcb versions.

However, we can avoid this specific situation by just checking whether
an abstract socket exists for a given display number, and choosing a
higher number if so.