fix(mmserver): increment stream_seq when a refresh packet is sent
This logically makes sense, and it makes it easier for clients to reset
their decoder. It requires moving sequence number generation further down
the pipeline.

This also contains a minor change to concatenate headers with the
keyframe packet. We were sending headers as their own packet, which isn't
standard and is a bit inefficient.