feat!: send hierarchical_layer as video frame metadata
The motivation for this (breaking) change is that the `frame_optional`
field doesn't provide enough information to the decoder on all
platforms. In particular, VideoToolbox on Mac, when confronted with a
dropped frame, breaks if you send it any frames depending on it until
it's recovered. If there are more than two layers, the middle layer
isn't exactly optional, but it's only necessary to exclude higher layers
until we recover.