fix(mmserver): allow waiting on one frame while encoding another
This significantly improves performance, almost embarassingly so. The
key insight is that the encode operation submitted to the GPU will wait
on the composite operation, which itself waits on the game to finish
rendering; so if a game is slower than our virtual framerate, we end up
backing up very quickly.