fix(mmserver): handle EAGAIN during sendmmsg
This is not actually sufficient for error handling in sendmmsg. The
following cases are unconsidered:

 - Only some of the buffer are sent
 - Only some of one of the iovecs is sent

We'll handle this when we migrate to rustix's sendmmsg, which is waiting
on the following PRs:

 - https://github.com/bytecodealliance/rustix/pull/1171
 - https://github.com/bytecodealliance/rustix/pull/1409

Note that those cases should be gracefully-ish handled by FEC and the
protocol (they're equivalent to dropped or biffed packets).